backend/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
boto3==1.34.10
Werkzeug==3.0.1
